[Decomp][ObjectMgr] Cohesion split: extract creature loaders to ObjectMgrCreatures.cpp
